Add files via upload
Right now every pizza is set to Hawaiian. There is still a lot of work to be done. Check the value pizza range. Clicking on some of the "+" will add a pizza to a tally and a bar will appear. I am actively working on making this appear more like our mock ups.